requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
openpyxl==3.1.2
requests==2.31.0
pyyaml==6.0.1
tqdm==4.66.2
loguru==0.7.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
Pillow==10.2.0
google-api-python-client==2.114.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.18.0
gspread==5.12.4
google-cloud-storage==2.14.0
selenium==4.18.1
webdriver-manager==4.0.1
openai>=1.50.0
anytree==2.10.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.2.0
flake8==7.0.0
mypy==1.7.1
isort==5.13.2
ruff==0.1.15